# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "\u003cadd description here\u003e",
        "title": "ORCHID_FOOTWARES",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:9080",
    "basePath": "/",
    "paths": {
...
        "/admin/salesreport": {
            "get": {
                "description": "Admin can download sales report from here",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin",
                    "sales report"
                ],
                "summary": "sales report",
                "operationId": "sales report",
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            }
        },
...
        "/admin/superadminlogin": {
            "post": {
                "description": "Admin can sign up with email and password",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin",
                    "home"
                ],
                "summary": "Admin Login",
                "operationId": "Admin-login",
                "parameters": [
                    {
                        "description": "Login credentials",
                        "name": "login",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.AdminLoginResponse"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            }
        },
...
        "/user/addaddress": {
            "post": {
                "description": "User can add their address here...",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User",
                    "address management"
                ],
                "summary": "Add Address",
                "operationId": "add-address",
                "parameters": [
                    {
                        "description": "Address data",
                        "name": "AddAddressResponse",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.AddAddressResponse"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            }
        },
        "/user/addcart": {
            "post": {
                "description": "User can add items to their cart",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User",
                    "cart management"
                ],
                "summary": "Add Cart",
                "operationId": "add-cart",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Product_ID",
                        "name": "product_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Product_Quantity",
                        "name": "product_quantity",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            }
        },
...
        "/user/deleteaddress": {
            "delete": {
                "description": "Delete address of a user",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User",
                    "address management"
                ],
                "summary": "Delete address",
                "operationId": "delete-address",
                "parameters": [
                    {
                        "description": "Idelete addressD ",
                        "name": "AddResponse",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.AddResponse"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            }
        },
        "/user/deleteitem": {
            "delete": {
                "description": "User delete cart items",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User",
                    "cart management"
                ],
                "summary": "delete-cart items",
                "operationId": "delete-cart",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id",
                        "name": "productId",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "quantity",
                        "name": "quantity",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            }
        },
...
        "/user/editaddress": {
            "patch": {
                "description": "User can change their address here ...",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User",
                    "address management"
                ],
                "summary": "Edit Address",
                "operationId": "edit-address",
                "parameters": [
                    {
                        "description": "information",
                        "name": "EditAddressResponse",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/handlers.EditAddressResponse"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            }
        },
...
        "/user/viewaddress": {
            "get": {
                "description": "User can view their address",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User",
                    "address management"
                ],
                "summary": "viewaddress",
                "operationId": "view-address",
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            }
        },
        "/user/viewcart": {
            "get": {
                "description": "User can view their cart",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User",
                    "cart management"
                ],
                "summary": "View Cart",
                "operationId": "view-cart",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Page ",
                        "name": "page",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "perpage",
                        "name": "perpage",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            }
        },
...
    },
    "definitions": {
        "handlers.AddAddressResponse": {
            "type": "object",
            "properties": {
                "district": {
                    "type": "string"
                },
                "house_name": {
                    "type": "string"
                },
                "pin_Number": {
                    "type": "integer"
                },
                "place": {
                    "type": "string"
                },
                "state": {
                    "type": "string"
                }
            }
        },
        "handlers.AddResponse": {
            "type": "object",
            "required": [
                "address_id"
            ],
            "properties": {
                "address_id": {
                    "type": "integer"
                }
            }
        },
        "handlers.AdminLoginResponse": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
        "handlers.EditAddressResponse": {
            "type": "object",
            "properties": {
                "district": {
                    "type": "string"
                },
                "house_name": {
                    "type": "string"
                },
                "pin_Number": {
                    "type": "integer"
                },
                "place": {
                    "type": "string"
                },
                "state": {
                    "type": "string"
                }
            }
        },
...
    }
}